add job search with optional parameters

+++ src/main/java/core/basesyntax/dreamjob/dto/job/internal/BatchOfJobsResponseDto.java
@@ -2,5 +2,5 @@ package core.basesyntax.dreamjob.dto.job.internal;
 
 import java.util.List;
 
-public record BatchOfJobsResponseDto(int count, List<JobsShortenedResponseDto> jobDtos) {
+public record BatchOfJobsResponseDto(long count, List<JobsShortenedResponseDto> jobDtos) {
 }
